update weight, delete exercise buttons added
update weight, delete exercise buttons added. Delete button is
functional. The buttons use modals, where users have to enter the ID,
in order to delete or update an exercise.

Thinking about using radio buttons in conjunction with the delete or
update buttons. Maybe use Jquery to grab the value of an exercises ID
then feed this into one of the functions that pass data to the api...
have to think on this more..